Finish execution instructions. Update script files.
