--- test/run_test.bat
@@ -1,9 +1,0 @@
-@echo off
-:: Enter test folder
-cd AutoTest
-
-:: install node modules
-npm ci
-
-:: run tests
-npm run AllTest
